[CustomOp] Update thresholding new class hierarchy
